+++ crates/iota-source-validation/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+name = "iota-source-validation"
+version.workspace = true
+authors = ["IOTA Foundation <contact@iota.org>"]
+edition = "2021"
+license = "Apache-2.0"
+publish = false
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+colored.workspace = true
+futures.workspace = true
+thiserror.workspace = true
+tracing.workspace = true
+
+iota-json-rpc-types.workspace = true
+iota-move-build.workspace = true
+iota-package-management.workspace = true
+iota-sdk.workspace = true
+iota-types.workspace = true
+
+move-binary-format.workspace = true
+move-bytecode-source-map.workspace = true
+move-command-line-common.workspace = true
+move-compiler.workspace = true
+move-core-types.workspace = true
+move-package.workspace = true
+move-symbol-pool.workspace = true
+
+flate2.workspace = true
+tar.workspace = true
+tempfile.workspace = true
+ureq.workspace = true
+
+[dev-dependencies]
+expect-test.workspace = true
+rand.workspace = true
+tempfile.workspace = true
+tokio = { workspace = true, features = ["macros", "test-util"] }
+tracing.workspace = true
+
+iota-test-transaction-builder.workspace = true
+test-cluster.workspace = true
