+++ crates/iota-source-validation/fixture/c/sources/c.move
@@ -1,0 +1,9 @@
+// Copyright (c) Mysten Labs, Inc.
+// Modifications Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+module c::c {
+    public fun c(): u64 {
+        43
+    }
+}
