+++ crates/iota-source-validation/fixture/c/sources/d.move
@@ -1,0 +1,9 @@
+// Copyright (c) Mysten Labs, Inc.
+// Modifications Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+module c::d {
+    public fun d(): u64 {
+        44
+    }
+}
