+++ crates/iota-source-validation/fixture/versioned-a-depends-on-b/sources/a.move
@@ -1,0 +1,12 @@
+// Copyright (c) Mysten Labs, Inc.
+// Modifications Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+module a::a {
+    use b::b::b;
+    use b::b::c;
+    
+    public fun a() : u64 {
+        b() + c()
+    }
+}
