+++ crates/iota-source-validation/fixture/versioned-b/sources/b.move
@@ -1,0 +1,13 @@
+// Copyright (c) Mysten Labs, Inc.
+// Modifications Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+module b::b {
+    public fun b(): u64 {
+        42
+    }
+
+    public fun c(): u64 {
+        b() + 1
+    }
+}
