+++ crates/iota-source-validation/fixture/z/Move.toml
@@ -1,0 +1,10 @@
+[package]
+name = "z"
+version = "0.0.1"
+published-at = "$STORAGE_ID"
+
+[dependencies]
+Iota = { local = "$REPO_ROOT/crates/iota-framework/packages/iota-framework" }
+
+[addresses]
+z = "$RUNTIME_ID"
