--- crates/sui-source-validation/Cargo.toml
@@ -1,47 +1,0 @@
-[package]
-name = "sui-source-validation"
-version.workspace = true
-authors = ["Mysten Labs <build@mystenlabs.com>"]
-license = "Apache-2.0"
-publish = false
-edition = "2021"
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-anyhow.workspace = true
-colored.workspace = true
-thiserror.workspace = true
-tracing.workspace = true
-futures.workspace = true
-
-sui-json-rpc-types.workspace = true
-sui-types.workspace = true
-sui-sdk.workspace = true
-sui-move-build.workspace = true
-
-move-binary-format.workspace = true
-move-bytecode-source-map.workspace = true
-move-command-line-common.workspace = true
-move-compiler.workspace = true
-move-core-types.workspace = true
-move-package.workspace = true
-move-symbol-pool.workspace = true
-
-tar.workspace = true
-tempfile.workspace = true
-flate2.workspace = true
-ureq.workspace = true
-
-
-[dev-dependencies]
-
-expect-test.workspace = true
-rand.workspace = true
-tempfile.workspace = true
-tokio = { workspace = true, features = ["macros", "test-util"] }
-tracing.workspace = true
-
-sui-test-transaction-builder.workspace = true
-test-cluster.workspace = true
