--- crates/sui-source-validation/fixture/b/sources/b.move
@@ -1,8 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-module b::b {
-    public fun b(): u64 {
-        42
-    }
-}
