--- crates/sui-source-validation/fixture/c/sources/b.move
@@ -1,8 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-module c::b {
-    public fun b(): u64 {
-        42
-    }
-}
