--- crates/sui-source-validation/fixture/c/sources/c.move
@@ -1,8 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-module c::c {
-    public fun c(): u64 {
-        43
-    }
-}
