--- crates/sui-source-validation/fixture/c/sources/d.move
@@ -1,8 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-module c::d {
-    public fun d(): u64 {
-        44
-    }
-}
