--- crates/sui-source-validation/fixture/versioned-b/Move.lock
@@ -1,11 +1,0 @@
-# @generated by Move, please check-in and do not edit manually.
-
-[move]
-version = 0
-manifest_digest = "C70E0438875A0C3B10FA5995F317F701134E629D1B2E5F3DCA3621E50034F809"
-deps_digest = "E3B0C44298FC1C149AFBF4C8996FB92427AE41E4649B934CA495991B7852B855"
-
-[move.toolchain-version]
-compiler-version = "$COMPILER_VERSION"
-edition = "legacy"
-flavor = "sui"
