--- crates/sui-source-validation/fixture/versioned-b/sources/c.move
@@ -1,8 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-module b::c {
-    public fun c(): u64 {
-        43
-    }
-}
