--- crates/sui-source-validation/fixture/versioned-b/sources/d.move
@@ -1,8 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-module b::d {
-    public fun d(): u64 {
-        44
-    }
-}
