--- crates/sui-source-validation/fixture/z/Move.toml
@@ -1,10 +1,0 @@
-[package]
-name = "z"
-version = "0.0.1"
-published-at = "$STORAGE_ID"
-
-[dependencies]
-Sui = { local = "$REPO_ROOT/crates/sui-framework/packages/sui-framework" }
-
-[addresses]
-z = "$RUNTIME_ID"
