--- crates/sui-source-validation/fixture/z/sources/b.move
@@ -1,8 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-module z::b {
-    public fun foo(x: u64): u64 {
-        x
-    }
-}
